breakout channel controller
